Export circuit class to another module
